Added required methods to some of the entities

+++ Warzone/src/main/java/com/warzone/Entities/Continent.java
@@ -22,7 +22,7 @@ public class Continent {
             return d_Name;
       }
 
-      public Long get_ControlValue() {
+      public int get_ControlValue() {
             return d_ControlValue;
       }
